requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
praw==7.7.1
requests==2.31.0
aiohttp==3.9.1
pillow==10.1.0
opencv-python==4.8.1.78
pandas==2.1.3
torch==2.1.0
timm==0.9.10
apscheduler==3.10.4
pyyaml==6.0.1
python-dotenv==1.0.0
boto3==1.34.0
pydantic==2.5.0

